Test a crear pago

+++ microservicio/dominio/src/main/java/com/ceiba/pago/servicio/impuesto/ImpuestoPredial.java
@@ -2,6 +2,6 @@ package com.ceiba.pago.servicio.impuesto;
 
 public interface ImpuestoPredial {
 
-    public double calcular();
+    public int calcular();
 
 }
